backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
pydantic==2.10.6
pydantic-settings==2.8.1
PyJWT==2.10.1
pytest==8.3.5
httpx==0.28.1
